agents/sentinel-agent/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.13.12
google-cloud-firestore==2.11.1

# Shared library (installed via Docker COPY)